Initial Zombie Wave System, plus more.
My initial foray into a wave system.

Many ideas are in place, but have yet to be tested.

Ideas include:
Difficulty modifier: Health and EXP increases with time.
Level modifier: Levels and SkillPoints are not earned linearly.

The turret can be placed in front of Zed the same way the bullet is
spawned in front of Zed.

The gameStateController is made to manage the wave system. ZombieWave is
also implemented as a container for ZombieSpawnJobs.

The *problem* is that Unity hangs when handleSpawnJobs is reached. I
suspect it has to do with the spawnJobs not being set correctly. I will
fix it with future update.