Zombie Wave System working.
The problem lay in my wave being referenced before it could be created.
Since Unity doesn't give any indications to which script is called
first, you have to enforce referencing to happen at the End of the
frame. So yeah, that's been fixed.

Skill points will increase in step with levels. Experience points have
been added too. I can continue to improve the leveling system with time,
but for now I think we have a good system going for us.