Updated survival zombie spawn sequence
The old version did not use ZombieWaves.

The new one does.

I recommend letting the ZombieSpawnEngine determine the victory state.
The Zombie Spawner is an object that controls a single spawn. The Spawn
Engine can check itself to see if it has any more waves, and the
GameStateController can check if there are still-undead zombies on the
field. It's best to let the GameStateController decide for itself when
victory state is achieved.