Updated rack obstacles, lv2 map.
Pushing Zed into the obstacle would cause the zombies to fail in
pathfinding and stand still.

The zed collision box on the rack has been expanded to work around this.
In addition, I turned the rack into a prefab for easy access.

I changed the rack layout of Level 2 as I was too lazy to mimic the
original layout when replacing the old racks with the new prefab. If the
layout is no good, please help me to change it back. Thanks.