ZombieMines!
ZombieMines level prototyped. The Main Menu has a button to enter this
level. It looks ugly and ham-fisted, but the concept is shown
sufficiently.

In addition, mines work properly-ish now. Still need to work out how the
radius can be better controlled. In the meantime the mines *seem* to
work as expected.

Sound effects have been worked in too, so we got that going for us,
which is nice.