Fixed the Lv1 victory condition problem
The cause was the BossZombie's spawns having a spawn time that was way
beyond the current time. Anyway it is fixed, both the boss not spawning
and the victory condition.

Lv2 should be fine too. We can move forward now.

Also, I've balanced the assault rifle. 3 hits to kill a standard zombie,
and fires about 40% faster