Fixed the Level 2 hang problem
Take note: TimeScale is set to zero on the main menu.

You have to set it back to one before loading the level again. That was
the problem. Fixed now.

I have also added boss footsteps to make him sound more giant-like.

Some cosmetic changes such as not actually forcing a  reload of the
weapons on game start.

I also want to change Time.time to Time.timeSinceLevelLoad, since they
will not be the same as level loading will not reset Time.time.

I encountered problems with spawning thanks to said fact. It would be
good to get everything else on the same page as well.