Quick attach of waveOverlay.
gameStateController decides when to plaster the waveOverlay text.

I'm not sure how to get it to animate properly. It might be easier to
control it if an onGUI function is used instead.

Unless, of course, making the mainCamera the parent of the waveOverlay
is not the way to go. I'm not sure how to go about this.